# swagger.yaml
openapi: 3.0.0
info:
  title: Bookky API ของเบส
  description: ระบบจัดการหนังสือสุดจ๊าบ มีระบบค้นหาและเพิ่มข้อมูล
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: เครื่องของฉันเอง (Localhost)

components:
  schemas:
    # แม่พิมพ์สำหรับตอนส่งข้อมูลเข้าไปเพิ่ม (POST, PUT)
    BookInput:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: "แฮร์รี่ พอตเตอร์"
        price:
          type: number
          example: 500

paths:
  # ----------------------------------------
  # จัดการข้อมูลทั้งหมด (/bookky)
  # ----------------------------------------
  /bookky:
    get:
      summary: ดึงข้อมูลหนังสือทั้งหมด
      parameters:
        - in: query
          name: sort
          schema:
            type: string
          description: พิมพ์คำว่า 'true' ถ้าอยากให้เรียงข้อมูล
      responses:
        '200':
          description: ดึงข้อมูลสำเร็จจร้า
    
    post:
      summary: เพิ่มหนังสือเล่มใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: เก่งมากเสร็จแล้ว (เพิ่มข้อมูลสำเร็จ)
        '400':
          description: ใส่ให้ถูกนะเว่ย (ส่งข้อมูลมาไม่ครบ)

  # ----------------------------------------
  # จัดการข้อมูลรายเล่ม (/bookky/{id})
  # ----------------------------------------
  /bookky/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: ระบุ ID ของหนังสือที่ต้องการจัดการ

    put:
      summary: แก้ไขข้อมูลหนังสือ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '200':
          description: Update toy complete
        '404':
          description: book not found

    delete:
      summary: ลบหนังสือทิ้ง
      responses:
        '200':
          description: Deleted book from inventory successfully
        '404':
          description: book not found

  # ----------------------------------------
  # ระบบค้นหา (/search)
  # ----------------------------------------
  /search:
    get:
      summary: ค้นหาหนังสือจากชื่อ
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: พิมพ์ชื่อหนังสือที่ต้องการหา
      responses:
        '200':
          description: เจอแล้วจร้า